ui/requirements.txt
selenium
pytest-bdd
pytest-print